Parse starttime as signed, so an out-of-range value is refused not wrapped (#19)
CodeQL go/incorrect-integer-conversion, high, internal/collect/linux.go:295.

starttime came from strconv.ParseUint and was then converted with int64() to be
added to the boot time. A value above MaxInt64 becomes negative in that
conversion and raises nothing: 2^63 renders as -2922768221-05-18T01:37:22Z, which
is measured rather than reasoned about. On a page whose entire job is to say when
a process started, a fabricated date is worse than an absent one.

Fixed at the root rather than guarded at the use site: StartTicks is parsed with
ParseInt and typed int64, which removes the conversion instead of bounds-checking
it. /proc renders the field unsigned, so an implausible value is now refused where
it is read, and the caller omits the start time exactly as it already does when
/proc/stat cannot be read at all.

CPUTicks stays uint64 deliberately. It is only ever compared as a delta against a
previous reading, and unsigned is the correct type for a counter that is checked
for going backwards.

The test asserts both directions: 2^63 is refused, and MaxInt64 itself still
parses, so the guard cannot be off by one.

+++ internal/collect/linux.go
@@ -292,7 +292,7 @@ func readProcesses(opts Options) ([]Process, error) {
 			// wrong when /proc/stat could not be read: a start time computed from a boot time we do not
 			// have would be an arbitrary date, and a date is believed in a way a dash is not.
 			if bootedAt > 0 {
-				at := time.Unix(bootedAt+int64(times.StartTicks)/userHz, 0).UTC()
+				at := time.Unix(bootedAt+times.StartTicks/userHz, 0).UTC()
 				p.StartedAt = &at
 			}
 		}
